// src_backup_20250808_181247/components/fahndungskarte/types.ts
export type CategoryType =
  | "WANTED_PERSON"
  | "MISSING_PERSON"
  | "UNKNOWN_DEAD"
  | "STOLEN_GOODS";

export type PriorityType = "urgent" | "new" | "normal";

export interface FahndungsData {
  step1: {
    title: string;
    category: CategoryType;
    caseNumber: string;
  };
  step2: {
    shortDescription: string;
    description: string;
    priority: PriorityType;
    tags: string[];
    features: string;
  };
  step3: {
    mainImage: string;
    mainImageUrl?: string; // URL des hochgeladenen Hauptbildes
    additionalImages: string[];
    additionalImageUrls?: string[]; // URLs der hochgeladenen zusätzlichen Bilder
  };
  step4: {
    mainLocation?: { address: string };
  };
  step5: {
    contactPerson: string;
    contactPhone: string;
    contactEmail?: string;
    department: string;
    availableHours: string;
  };
}